requirements.txt
PyPDF2
Pillow
reportlab
tk
PyMuPDF